Allow 'user' to be specified as a join() argument and as an argument to setDailyLang().
In the process, do a bit of cleanup in our language-handling logic, removing an unnecessary mapping between languages and UI languages (since we're doing language-prefix-matching instead), and remove 'en-us' since it's not doing anything for us and isn't documented as supported language anyway.

If someday down the road we outgrow a simple prefix-matching strategy (e.g. 'fr-ca' --> 'fr'), we can revisit this.