Expose token configuration through `room()` method, matching how we expose room and domain configuration.
This is the simplest way of unblocking the new prebuilt UI from taking token-provided configuration into account when making decisions, and is at least consistent with our room and domain configuration pattern. However, I think it's becoming clear that we could really benefit from a more ergonomic and consistent approach to configuration going forward.

Today, if an app wants to know which configuration values are actually in effect for the current meeting, they might need to look at domain, room, and token values, and know which take precedence over which. There *are* ways of looking up specific aspects of meeting configuration that are currently in effect—for example, `getDailyLang()` or `participants().local.owner`—but there are lots of gaps. If we wanted to add the ability to look up which value is in effect for any configuration property, we have no pattern to follow other than adding dedicated getter methods for every single property. This would quickly clutter up our API.

Furthermore, there are various client-side-only-settable bits of configuration like whether to show the participants bar or local video, which beg the question why there aren't domain- or room- or token-settable equivalents.

Anyway, all that is to say: I think there's an elegant solution out there waiting to be designed and implemented, but let's start with at least making the information available, which this PR does.