Make it possible for customers to use room() as soon as it's ready (i.e. as soon as it can successfully fetch room info from the server) by accessing it in an 'access-state-updated' handler:
callObject.on('access-state-updated', async () => {
  let roomInfo = await callObject.room();
  if (roomInfo.id) {
    console.log('room info is available!', roomInfo);
  }
});

Prior to this change, there was no event correlated to room() being ready for use—the best you could do is just poll for its availability.

Note that accessState() is a bit wonky in a couple of ways: it doesn't go to 'none' when the user is rejected during knocking—instead, the meeting will get into an 'error' state. Also note that accessState() is not cleared on leave(), only on destroy(). For those reasons, the simplest thing is to do what the code block above does and check for `roomInfo.id` to know that the room info is in fact available (otherwise, you'd have to indirectly check both accessState() and meetingState(), which is more complicated).