Add 'available-devices-updated' even on daily-js, with a similar implementation as react-native-daily-js.
Note that we could've instead chosen to piggyback on the default-device-change polling mechanism, but the approach taken here is more straightforward from a code standpoint, as it requires no new state to keep track of.

Testing forthcoming to validate that using the browser's devicechange event means we can get away without polling.

If so, then in the future we should consider replacing our own default-device-change polling mechanism with a devicechange-event-based mechanism.