Introduce a `callObjectBundleUrlOverride` field under `dailyConfig` in the `DailyCallOptions`.
This makes it possible for customers to host the call machine bundle themselves to avoid having to specify `script-src https://*.daily.co` in their CSP (if they’re using `avoidEval: true` already).

This is an **advanced** configuration, as using it puts the onus on the customer to ensure that they’re serving up the right up-to-date bundle. (Note that it will actually soon become slightly easier for customers to be confident that they’re serving the right version of the call machine bundle, once we finish the work of “pinning” each version of daily-js to a specific call machine bundle version.)

This change has been validated with `callObjectBundleUrlOverride` pointing to the bundle at the same host as well as an arbitrary host, with a CSP configured accordingly.