Api to allow setting the idleTimeOut.
When user switch to other tab some browsers throttle the CPU and
cause stream to pause. becaues of IdleTimeOut logic on worker, livestream
or recording is automatically stopped.
With this API we are allowing user to override this behaviour and let
stream run as long as they want.

Note: Stream may take 2-3 mins more than the set value.