Add support for new user data API
This commit implements support for setting a new `userData` field
on the local user that is shared through presence data to remote
users. userData can also be set via a frame property.

See the userData spec for more details (note: this does not
implement the entire spec. more to come...):

https://www.notion.so/dailyco/User-Data-db4c4e14f87242c4ae8c44349482bc32

a meager start

about there with just the basic API and passing of data