feat(daily-js): Implement an optional `reason` field in the payload of the `'participant-left'` event, and a first reason `'became-hidden'`, which occurs when a participant goes from having presence to being hidden. The only way this occurs today is if they have their `hasPresence` permission revoked.
Note that the server doesn't stick the reason in the `msgData` field, so we can keep the convention of `msgData === null` signifying a participant disconnecting. Otherwise we would run into additional complexity, not least of which would be adding client-version-based branching in the server code to avoid breaking clients in the wild who are expecting `null` to signify "left".