feat(dj): Allow updating volume of RMP.
Added a optional `volume` parameter in `DailyRemoteMediaPlayerSettings`.
volume is optional parameter, with default value=1.0. which means
original volume of the file. value can range from 0.0 (mute) to
10.0 (max volume).

state is `DailyRemoteMediaPlayerSettings` is option now, which means
dont update the current state, if player is player it remain in playing
state.
`updateRemoteMediaPlayer` can update only volume, only state or both.
if `volume` or `state` is not speciifed player continue in the
current state.