Some error handling in the case where user data or meeting session data isn't structured-clonable and therefore not able to be sent to the call machine.
This can occur despite the data having undergone a stricter validation that it can be JSON serialized then deserialized without changing, because that validation only prints a warning and doesn't stop processing, since we want to be permissive in what we allow users to pass us.

An example of an argument to `setUserData()` that will print a "will be changed during serialization" warning but *won't* trigger a "not-structured-clonable" error is `{ foo: undefined }`.