fix(daily-js): rmpAudio and rmpVideo track state is wrong when paused.
When remote-media-player is paused, it is equivalent to camera/mic mute,
Hence the rmptTrack state should be `off`.
When rmp is paused, it is signalled from server and based on signalling
correct `track-stopped` and `participant-updated` events are generated.
similarly when rmp is playing again, track-started is generated with
`playable` state.

Note: when player is paused only the video's MediaStreamTrack receive
onMute callback, audio MST never receive onMute. It is not an issue as
we handle track state through server signalling.