fix(cm): Fix a problem where `setInterval()` intervals that the call object starts are never cleaned up if the user doesn't end up `join()`ing a call first. Intervals should be cleaned up when the user invokes `destroy()`, whether or not they ever `join()`ed.
The poll for default device changes is an example of one of these intervals. It may be triggered by `startCamera()` (and then unmuting your mic, if not already unmuted). Prior to this commit, invoking `destroy()` after `startCamera()` (without `join()`ing in between) will leave the poll running indefinitely.

This is a very longstanding issue.

Today, the `_exitMeetingCleanup()` function—which is only run if the user first `join()`s a call and then `leave()`s, or there is an error—is not called as part of `destroy()`. This commit changes that by ensuring that `destroy()` always runs `leave()` first. If the user has already left the call, no additional work is done.

There is one side effect of this change: now, after `destroy()` is run, the call object will always be left with `meetingState()` as either `'left-meeting'` or `'error'`, whereas previously the call object could also be left in a number of other states.