fix(dj): Fix bug where `getInputSettings()` would erroneously report after `load()` (or `startCamera()`, or `join()`) that you had no processors specified (even if you specified one in the preload cache via a constructor argument or a previous called to `updateInputSettings()`, for example), until the first `input-settings-updated` event fired.
This was making it so that the `bg-effect` cookie logic in prebuilt wasn't working as intended: it would *always* think the call object had no input settings previously specified and would try to apply its cookie-provided processor settings anyway. The logic *happened* to mostly work since the preload cache's processing setting would "win" because it got the "lock" and preventing the subsequent cookie-provided settings from taking effect. But it wouldn't work if the preload cache processor settings were of `type: 'none'`.