imp(cm): clean up the cruft around calculating and dispersing network threshold changes
there was a LOT of no-op, unnecessary work that mostly dates back to
supporting the old prebuilt-ui. this commit cleans up all the logic to remove
bits that are unused and rename variables and function to how they are
actually used. it essentially is a no-op except for adding logs for when a
threshold change occurs.