Rename SIP_* to DIALIN_* events and clean up event handling
We have three use cases using SIP:
- SIP dial in
- PSTN dial in
- Dial out to SIP or PSTN

For the first two, use DIALIN_* events and DIALOUT_* for the last.
We also drop the missed event and add a reason to STOPPED event if
dial out is being stopped because remote party did not attend the
call.

We introduce total of 8 events, DIALIN and DIALOUT being symmetric.
Ideally we would have 4 events named SIP_* since technically this is
all SIP but in the interest of customers not knowing about SIP at
all in the first place, introduce separate events for dial-in and
dial-out.