Drop instanceId & use SIP session id for the SIP/PSTN/DialOut participant
Summarising details from discussion on PR for the reasoning behind this
change. A lot of what follows is as succinctly described by Arun.

For recordings, it made sense to separate instanceId and recordingId,
since we didn't have the promise to tie each startRecording() call to
the corresponding recordingId, and we wanted to make sure that a user
is careful about triggering more than one at a time. So they provide
an instance ID themselves, and we use that in the API to manage each
separate recording/streaming session.

With dialout, we do have a promise in the response to the startDialout(),
so this is less of a concern, the client has a good way to tie each call
to the success/failure response, and thus can have a server-generated
handle for each dialout instance without having to worry about how the
client will know which response is for which call.

So the idea would be to drop the instanceId for dialout, and use the
generated sipSessionId as the single identifier i.e. sessionId.

We use the SIP sessionId for participant ID as well, for the caller
to know that this the participant is a SIP/PSTN/DialOut user.

For the completion of discussion's sake, note that

We return from the backend and complete the startDialOut() call when
the SFU has started a SIP worker, and that worker has kicked off SIP
"INVITE" to the number/URI provided. Effectively, we have "dialled"
the provided number/URI.

The actual setting up of the connection might take a few hops, then
it might ring for a while, it may or may not finally actually be
answered, and if it is, then we negotiate media with the remote end
and finally set up the participant when we're ready to produce/consume
audio/video.