Return to use of 'callFrameId' terminology...
callFrameId ends up re-routed and part of many of our external facing
daily-js events. If we rename to anything other than callFrameId, we'd
want to update ALL usage for consistency, but that could potentially
break customers, despite the fact that it should have zero use (and really
shouldn't be part of those events). But here we are...

If/when we DO update this variable, we should call it callClientId which
will be consistent with general terminology and will not conflict with
any other existing terminology.